refactor: auth and workspace
