+++ apps/electron/src/main/modules/auth/auth.service.ts
