+++ apps/electron/src/main/modules/workspace/workspace.ipc.ts
@@ -1,5 +1,5 @@
 import { ipcMain } from "electron";
-import { getWorkspaceService } from "@/main/modules/workspace/workspace-service";
+import { getWorkspaceService } from "@/main/modules/workspace/workspace.service";
 import type { WorkspaceCreateConfig } from "@mcp_router/shared";
 
 /**
