+++ apps/electron/src/main/modules/workspace/workspace.service.ts
